+++ src/config/index.js
@@ -1,11 +1,11 @@
 export const effects = [
-  { id: 'autoFilter', name: 'Auto filter' },
-  { id: 'autoPanner', name: 'Auto panner' },
-  { id: 'autoWah', name: 'Auto wah' },
-  { id: 'bitCrusher', name: 'Bit crusher' },
+  { id: 'autoFilter', name: 'Auto Filter' },
+  { id: 'autoPanner', name: 'Auto Panner' },
+  { id: 'autoWah', name: 'Auto Wah' },
+  { id: 'bitCrusher', name: 'Bit Crusher' },
   // { id: 'chorus', name: 'Chorus' },
   { id: 'distortion', name: 'Distortion' },
-  { id: 'feedbackDelay', name: 'Feedback delay' },
+  { id: 'feedbackDelay', name: 'Feedback Delay' },
   { id: 'freeverb', name: 'Freeverb' },
   { id: 'panVol', name: 'Volume/Pan' },
   // { id: 'reverb', name: 'Reverb' },
@@ -13,12 +13,14 @@ export const effects = [
 ];
 
 export const instruments = [
-  { id: 'AMSynth', name: 'AM synth' },
-  { id: 'duoSynth', name: 'Duo synth' },
-  { id: 'FMSynth', name: 'FM synth' },
-  { id: 'membraneSynth', name: 'Membrane synth' },
-  { id: 'monoSynth', name: 'Mono synth' },
-  // { id: 'polySynth', name: 'Poly synth' },
+  { id: 'amSynth', name: 'AM Synth' },
+  { id: 'duoSynth', name: 'Duo Synth' },
+  { id: 'fmSynth', name: 'FM Synth' },
+  { id: 'membraneSynth', name: 'Membrane Synth' },
+  { id: 'metalSynth', name: 'Metal Synth' },
+  { id: 'monoSynth', name: 'Mono Synth' },
+  // { id: 'noiseSynth', name: 'Noise Synth' }, // No sound, disabled for now
+  { id: 'pluckSynth', name: 'Pluck Synth' },
   { id: 'sampler', name: 'Sampler' },
   { id: 'synth', name: 'Synth' },
 ];