+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Upcoming] - 2020-01-08
 
 - Add `volume` and `isMuted` prop to `Song`
-- Add `polyphony` prop to `Instrument`
-- Add `oscillator` prop to `Instrument`
+- Add `polyphony` and `oscillator` prop to `Instrument`
+- Remove `type` of `polySynth` from `Instrument` as it is just a wrapper around other synths
+- Add polyphony for Instrument synths
 
 ## [0.2.2] - 2019-09-18